+++ src/components/common/ui/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import ChannelTalkBtn from './channel/ChannelTalkBtn';
 import Footer from './footer/Footer';
-import NavBar from './nav/NavBar';
+import NavBar from './header/NavBar';
 
 const Layout = () => {
   return (
