+++ src/components/common/ui/layout/header/NextNavBar.tsx
@@ -1,25 +1,23 @@
+'use client';
+
 import { useGetUserAdmin } from '@/api/user';
 import useActiveLink from '@/hooks/useActiveLink';
 import useActiveReportNav from '@/hooks/useActiveReportNav';
 import { useControlScroll } from '@/hooks/useControlScroll';
 import useScrollDirection from '@/hooks/useScrollDirection';
 import useAuthStore from '@/store/useAuthStore';
+import { usePathname } from 'next/navigation';
 import { useState } from 'react';
-import { useLocation } from 'react-router-dom';
-import GlobalNavItem from '../header/GlobalNavItem';
-import GlobalNavTopBar from '../header/GlobalNavTopBar';
-import NavOverlay from '../header/NavOverlay';
-import SideNavContainer from '../header/SideNavContainer';
-import SideNavItem from '../header/SideNavItem';
+import GlobalNavItem from './GlobalNavItem';
+import GlobalNavTopBar from './GlobalNavTopBar';
+import NavOverlay from './NavOverlay';
+import SideNavContainer from './SideNavContainer';
+import SideNavItem from './SideNavItem';
 
-const NavBar = () => {
-  const location = useLocation();
+const NextNavBar = () => {
+  const pathname = usePathname() ?? '';
 
   const [isOpen, setIsOpen] = useState(false);
-
-  const activeLink = useActiveLink(location.pathname);
-  const reportNavList = useActiveReportNav();
-  const scrollDirection = useScrollDirection(location.pathname);
 
   const { isLoggedIn } = useAuthStore();
 
@@ -28,10 +26,13 @@ const NavBar = () => {
     retry: 1,
   });
 
+  const activeLink = useActiveLink(pathname);
+  const reportNavList = useActiveReportNav();
+  const scrollDirection = useScrollDirection(pathname);
+
   const toggleMenu = () => {
     setIsOpen(!isOpen);
   };
-
   const closeMenu = () => {
     setIsOpen(false);
   };
@@ -47,9 +48,9 @@ const NavBar = () => {
       >
         {/* 1단 */}
         <GlobalNavTopBar
-          isNextRouter={false}
+          isNextRouter
           isActiveHome={activeLink === 'HOME'}
-          loginRedirect={encodeURIComponent(location.pathname)}
+          loginRedirect={encodeURIComponent(pathname)}
           toggleMenu={toggleMenu}
         />
         {/* 2단 */}
@@ -58,18 +59,18 @@ const NavBar = () => {
             <GlobalNavItem
               className="text-xsmall16"
               href="/program"
-              isNextRouter={false}
+              force
+              isNextRouter
               active={activeLink === 'PROGRAM'}
             >
               전체 프로그램
             </GlobalNavItem>
             <GlobalNavItem
               className="text-xsmall16"
-              isNextRouter={false}
+              isNextRouter
               active={activeLink === 'REPORT'}
               href={reportNavList.length === 0 ? '#' : reportNavList[0].href}
               subNavList={reportNavList}
-              force
             >
               서류 피드백 REPORT
             </GlobalNavItem>
@@ -82,18 +83,16 @@ const NavBar = () => {
             <GlobalNavItem
               className="text-xsmall16"
               href="/review"
-              isNextRouter={false}
+              isNextRouter
               active={activeLink === 'REVIEW'}
-              force
             >
               수강생 솔직 후기
             </GlobalNavItem>
             <GlobalNavItem
               className="text-xsmall16"
               href="/blog/list"
-              isNextRouter={false}
+              isNextRouter
               active={activeLink === 'BLOG'}
-              force
             >
               블로그
             </GlobalNavItem>
@@ -102,7 +101,8 @@ const NavBar = () => {
           <GlobalNavItem
             className="text-xsmall16"
             href="/about"
-            isNextRouter={false}
+            isNextRouter
+            force
             active={activeLink === 'ABOUT'}
           >
             렛츠커리어 스토리
@@ -110,35 +110,30 @@ const NavBar = () => {
         </nav>
       </div>
 
-      {/* 투명한 검정색 배경 */}
+      {/* 사이드 네비게이션 Overlay */}
       <NavOverlay isOpen={isOpen} onClose={closeMenu} />
 
       {/* 사이드 네비게이션 바 */}
-      <SideNavContainer
-        isNextRouter={false}
-        isOpen={isOpen}
-        onClose={closeMenu}
-      >
-        <SideNavItem href="/mypage/application" isNextRouter={false}>
+      <SideNavContainer isNextRouter isOpen={isOpen} onClose={closeMenu}>
+        <SideNavItem href="/mypage/application" isNextRouter force>
           마이페이지
         </SideNavItem>
         <hr className="h-1 bg-neutral-80" aria-hidden="true" />
-        <SideNavItem href="/about" isNextRouter={false}>
+        <SideNavItem href="/about" isNextRouter force>
           렛츠커리어 스토리
         </SideNavItem>
-        <SideNavItem href="/program" isNextRouter={false}>
+        <SideNavItem href="/program" isNextRouter force>
           프로그램
         </SideNavItem>
-        <SideNavItem href="/review" isNextRouter={false} force>
+        <SideNavItem href="/review" isNextRouter>
           수강생 솔직 후기
         </SideNavItem>
-        <SideNavItem href="/blog/list" isNextRouter={false} force>
+        <SideNavItem href="/blog/list" isNextRouter>
           블로그
         </SideNavItem>
         <SideNavItem
           href="/report/landing"
-          isNextRouter={false}
-          force
+          isNextRouter
           subNavList={reportNavList}
           onClick={(e) => e.stopPropagation()}
         >
@@ -146,14 +141,14 @@ const NavBar = () => {
         </SideNavItem>
         <hr className="h-1 bg-neutral-80" aria-hidden="true" />
         {isAdmin && (
-          <SideNavItem href="/admin" isNextRouter={false} force>
+          <SideNavItem href="/admin" isNextRouter force>
             관리자 페이지
           </SideNavItem>
         )}
         <SideNavItem
           className="notice_gnb"
           href="https://letscareer.oopy.io"
-          isNextRouter={false}
+          isNextRouter
           target="_blank"
           rel="noopener noreferrer"
         >
@@ -162,7 +157,7 @@ const NavBar = () => {
         <SideNavItem
           className="q&a_gnb"
           href="https://letscareer.oopy.io"
-          isNextRouter={false}
+          isNextRouter
           target="_blank"
           rel="noopener noreferrer"
         >
@@ -171,9 +166,9 @@ const NavBar = () => {
       </SideNavContainer>
 
       {/* 네비게이션 바 공간 차지 */}
-      <div className="h-[3.75rem] md:h-[111px]" />
+      <div className="h-[3.75rem] md:h-[111px]" aria-hidden="true" />
     </header>
   );
 };
 
-export default NavBar;
+export default NextNavBar;
